+++ src/plugins/xml/index.ts
@@ -2,11 +2,13 @@ import { javascriptProvider } from '../../io';
 import * as models from '../../models';
 import { provideAssertValueXPath } from './provideAssertValueXPath';
 import { xmlResponseInterceptor } from './xmlResponseInterceptor';
+import { parseXpathNamespace } from './xpathNamespaceHttpRegionParser';
 import { xpathVariableReplacer } from './xpathVariableReplacer';
 import * as xmldom from '@xmldom/xmldom';
 import * as xpath from 'xpath';
 
 export function registerXmlPuglin(api: models.HttpyacHooksApi) {
+  api.hooks.parse.addHook('xpath_ns', parseXpathNamespace, { before: ['variable'] });
   api.hooks.onResponse.addInterceptor(xmlResponseInterceptor);
   api.hooks.provideAssertValue.addHook('xpath', provideAssertValueXPath);
   api.hooks.replaceVariable.addHook('xpath', xpathVariableReplacer, { before: ['javascript'] });