+++ src/plugins/xml/xpathNamespaceHttpRegionParser.ts
@@ -1,0 +1,63 @@
+import * as models from '../../models';
+import { XPathProcessorContext } from './xpathProcessorContext';
+
+export async function parseXpathNamespace(
+  getLineReader: models.getHttpLineGenerator,
+  { httpRegion }: models.ParserContext
+): Promise<models.HttpRegionParserResult> {
+  const lineReader = getLineReader();
+  const next = lineReader.next();
+  if (!next.done) {
+    const textLine = next.value.textLine;
+
+    const match = /^\s*@xpath_ns\s+(?<key>[^\s=:]*)\s*=\s*"?(?<value>.*)"?\s*$/u.exec(textLine);
+
+    if (match && match.groups && match.groups.key && match.groups.value) {
+      const key = match.groups.key;
+      const value = match.groups.value;
+      httpRegion.hooks.execute.addHook(`xpath_ns_${key}`, async (context: XPathProcessorContext) => {
+        if (!context.options.xpath_namespaces) {
+          context.options.xpath_namespaces = {};
+        }
+        context.options.xpath_namespaces[key] = value;
+        return true;
+      });
+
+      return {
+        nextParserLine: next.value.line,
+        symbols: [
+          {
+            name: match.groups.key,
+            description: match.groups.value,
+            kind: models.HttpSymbolKind.variableDefinition,
+            startLine: next.value.line,
+            startOffset: 0,
+            endLine: next.value.line,
+            endOffset: next.value.textLine.length,
+            children: [
+              {
+                name: match.groups.key,
+                description: 'key',
+                kind: models.HttpSymbolKind.key,
+                startLine: next.value.line,
+                startOffset: next.value.textLine.indexOf(match.groups.key),
+                endLine: next.value.line,
+                endOffset: next.value.textLine.indexOf(match.groups.key) + match.groups.key.length,
+              },
+              {
+                name: match.groups.value,
+                description: 'value',
+                kind: models.HttpSymbolKind.value,
+                startLine: next.value.line,
+                startOffset: next.value.textLine.indexOf(match.groups.value),
+                endLine: next.value.line,
+                endOffset: next.value.textLine.indexOf(match.groups.value) + match.groups.value.length,
+              },
+            ],
+          },
+        ],
+      };
+    }
+  }
+  return false;
+}