+++ src/plugins/xml/xpathProcessorContext.ts
@@ -1,0 +1,7 @@
+import { ProcessorContext } from '../../models';
+
+export interface XPathProcessorContext extends ProcessorContext {
+  options: {
+    xpath_namespaces?: Record<string, string>;
+  };
+}